Add conditional compilation of Il2CppDumper
Moved the static data values to a json file. Adds two new build configurations: Debug_Dev and Release_Dev, which always generate new data if the version number has changed. This is saved to the system temp folder (Users/you/AppData/Temp/datavalues.json). Copy this to the core project folder to build a new version. The Debug/Release builds do not link Il2CppDumper and will fail if the json file is out of date.