Automatic generation of datavalues.json resource file.
Datavalues.json is now saved to AppData/Roaming/RaidExtractor. The dev version will look here after trying the inbuilt resource, and will use it if the version is correct.

The project has a new pre build step, which copies that file, if it exists, over the data file used for resource generation, so running the dev version, followed by building the non dev version is sufficient for updating the data.